3단계 - 로또(2등)
- 2nd 리뷰 반영
