+++ src/main/java/Main.java
@@ -1,8 +1,8 @@
-import domain.*;
-import factory.LottoFactory;
-import factory.RandomFactory;
-import view.InputView;
-import view.ResultView;
+import lotto.domain.*;
+import lotto.factory.LottoFactory;
+import lotto.factory.RandomFactory;
+import lotto.view.InputView;
+import lotto.view.ResultView;
 
 import java.util.Random;
 
@@ -16,8 +16,6 @@ public class Main {
 
     Lottos lottos = lottoFactory.randomLottos(count);
     ResultView.lotteryList(lottos);
-
-    Lottery lottery = InputView.lottery();
-    ResultView.stats(lottos.results(lottery));
+    ResultView.stats(lottos.results(InputView.lottery()));
   }
 }
