--- src/main/java/domain/BonusBallCondition.java
@@ -1,8 +1,0 @@
-package domain;
-
-public enum BonusBallCondition {
-  TRUE,
-  FALSE,
-  DONT_CARE
-}
-
