+++ src/main/java/lotto/domain/ConditionDto.java
@@ -1,0 +1,19 @@
+package lotto.domain;
+
+public class ConditionDto {
+  private int matchCount;
+  private boolean bonusBallCondition;
+
+  public ConditionDto(final Condition condition) {
+    this.matchCount = condition.matchCount();
+    this.bonusBallCondition = condition.bonusBallCondition();
+  }
+
+  public int getMatchCount() {
+    return this.matchCount;
+  }
+
+  public boolean getBonusBallCondition() {
+    return this.bonusBallCondition;
+  }
+}
